Add files via upload
Simple python script to get your pricing for your Steam market items. You would have to enter the URLs to get the JSON and the price when you have bought the specific item.
For my case, I already filled the variables in the script. The values are given in Euro. 
I already tried to do a live chart but I failed because of the amount of request which Steam is not allowing.